Refactor to separate string and buffer interfaces
