+++ lib/data/async.js
@@ -7,99 +7,41 @@ var exec = require( 'child_process' ).exec;
 var resolve = require( 'path' ).resolve;
 var join = require( 'path' ).join;
 var uuid = require( 'node-uuid' ).v4;
-var copy = require( 'utils-copy' );
 var cwd = require( 'utils-cwd' );
-var typeName = require( 'type-name' );
 var readFile = require( 'utils-fs-read-file' );
 var unlink = require( 'fs' ).unlink;
 var writeFile = require( 'fs' ).writeFile;
 var tmp = require( 'utils-tmpdir' )();
-var defaults = require( './../defaults.json' );
-var validate = require( './../validate.js' );
 var cmd = require( './../cmd.js' );
-
-
-// VARIABLES //
-
-var interfaces = {
-	'string,string,Object,function': [ 0, 1, 2, 3 ],
-	'string,Buffer,Object,function': [ 0, 1, 2, 3 ],
-	'string,Uint8Array,Object,function': [ 0, 1, 2, 3 ],
-	'string,string,function': [ 0, 1, -1, 2 ],
-	'string,Buffer,function': [ 0, 1, -1, 2 ],
-	'string,Uint8Array,function': [ 0, 1, -1, 2 ],
-	'string,Object,function': [ -1, 0, 1, 2 ],
-	'Buffer,Object,function': [ -1, 0, 1, 2 ],
-	'Uint8Array,Object,function': [ -1, 0, 1, 2 ],
-	'string,function': [ -1, 0, -1, 1 ],
-	'Buffer,function': [ -1, 0, -1, 1 ],
-	'Uint8Array,function': [ -1, 0, -1, 1 ]
-};
 
 
 // ASYNC //
 
 /**
-* FUNCTION: async( [dest,] data,[ options,] clbk )
+* FUNCTION: async( dest, data, opts, clbk )
 *	Asynchronously converts Markdown to reStructuredText.
 *
-* @param {String} [dest] - output reStructuredText file path
+* @param {String|Void} dest - output reStructuredText file path
 * @param {String|Buffer} data - Markdown data to convert
-* @param {Object} [options] - function options
-* @param {String} [options.flavor="gfm"] - Markdown flavor
+* @param {Object} opts - function options
+* @param {String} opts.flavor - Markdown flavor
 * @param {Function} clbk - callback to invoke after performing conversion
 * @returns {Void}
 */
-function async() {
+function async( dest, data, opts, clbk ) {
 	var outFile;
-	var options;
 	var inFile;
-	var types;
-	var dest;
-	var opts;
-	var clbk;
-	var data;
-	var err;
 	var dir;
 	var str;
-	var idx;
-	var len;
-	var i;
 
-	len = arguments.length;
-	types = new Array( len );
-	for ( i = 0; i < len; i++ ) {
-		types[ i ] = typeName( arguments[ i ] );
-	}
-	types = types.join( ',' );
-	idx = interfaces[ types ];
-	if ( idx === void 0 ) {
-		throw new Error( 'invalid input argument(s). No implementation matching `f('+types+')`.' );
-	}
-	if ( idx[ 0 ] >= 0 ) {
-		dest = arguments[ idx[0] ];
-	}
-	data = arguments[ idx[1] ];
-	if ( idx[ 2 ] >= 0 ) {
-		options = arguments[ idx[2] ];
-	}
-	clbk = arguments[ idx[3] ];
-
-	opts = copy( defaults );
-	if ( options ) {
-		err = validate( opts, options );
-		if ( err ) {
-			throw err;
-		}
-	}
 	dir = cwd();
 	debug( 'Current working directory: %s', dir );
 
-	inFile = join( tmp, uuid() );
+	inFile = join( tmp, uuid()+'.md' );
 	debug( 'Input file: %s', inFile );
 
 	if ( dest === void 0 ) {
-		outFile = join( tmp, uuid() );
+		outFile = join( tmp, uuid()+'.rst' );
 	} else {
 		outFile = resolve( dir, dest );
 	}
