+++ lib/index.js
@@ -4,7 +4,7 @@
 
 module.exports = require( './file/async.js' );
 module.exports.sync = require( './file/sync.js' );
-module.exports.fromString = require( './data/async.js' );
-module.exports.fromStringSync = require( './data/sync.js' );
-module.exports.fromBuffer = require( './data/async.js' );
-module.exports.fromBufferSync = require( './data/sync.js' );
+module.exports.fromString = require( './string/async.js' );
+module.exports.fromStringSync = require( './string/sync.js' );
+module.exports.fromBuffer = require( './buffer/async.js' );
+module.exports.fromBufferSync = require( './buffer/sync.js' );
